+++ 2021/Business/Day2/Service.cs
@@ -12,13 +12,13 @@
         public object RunTask1()
         {
             var simplePosition = _submarine.MoveSimple();
-            return simplePosition.depth * simplePosition.forward;
+            return (long)( simplePosition.depth * simplePosition.forward);
         }
 
         public object RunTask2()
         {
             var aimedPosition = _submarine.MoveWithAiming();
-            return aimedPosition.depth * aimedPosition.forward;
+            return (long)(aimedPosition.depth * aimedPosition.forward);
         }
     }
 }
